Rename hashing -> hash
